requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
sentence-transformers
huggingface-hub
# pip install torch --index-url https://download.pytorch.org/whl/cpu

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# streamlit
streamlit

# validation schema
pydantic
# pip install "pydantic[email]"

#chain graph
grandalf

#rag - document loader
langchain-community
pypdf